+++ CHANGELOG.md
@@ -7,9 +7,32 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ## [Unreleased]
 
+## [0.13.0] (2025-09-30)
+
+### Added
+
+- Add support for indexing and slicing 2D datasets
+  ([#48](https://github.com/MAAP-Project/gedi-subsetter/issues/48))
+
+### Removed
+
+- Remove use of `returns` library
+  ([#3](https://github.com/MAAP-Project/gedi-subsetter/issues/3))
+
+## [0.12.0] (2025-07-03)
+
+### Changed
+
+- Add "requester pays" flag for reading granule data files from S3
+  ([#138](https://github.com/MAAP-Project/gedi-subsetter/issues/138))
+- Add `tolerated_failure_percentage` input to control percentage of individual
+  granule failures to tolerate before failing a job.  Default tolerance is 0
+  (i.e., fail fast), thus any single failure will immediately fail the job.
+  ([#62](https://github.com/MAAP-Project/gedi-subsetter/issues/62))
+
 ## [0.11.0] (2025-04-25)
 
-## Changed
+### Changed
 
 - In addition to a logical name or an official DOI name, the value for the `doi`
   may now be a collection concept ID to uniquely specify a collection, because
@@ -247,4 +270,6 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.9.0]: https://github.com/MAAP-Project/gedi-subsetter/releases/tag/0.9.0
 [0.10.0]: https://github.com/MAAP-Project/gedi-subsetter/releases/tag/0.10.0
 [0.11.0]: https://github.com/MAAP-Project/gedi-subsetter/releases/tag/0.11.0
-[Unreleased]: https://github.com/MAAP-Project/gedi-subsetter/compare/0.11.0...HEAD
+[0.12.0]: https://github.com/MAAP-Project/gedi-subsetter/releases/tag/0.12.0
+[0.13.0]: https://github.com/MAAP-Project/gedi-subsetter/releases/tag/0.13.0
+[Unreleased]: https://github.com/MAAP-Project/gedi-subsetter/compare/0.13.0...HEAD